Bootstrap styling and added ejs view for wrong user

+++ controllers/teapots.js
@@ -5,44 +5,54 @@ const router      = express.Router();
 // models
 const Teapot      = require('../models/teapots.js');
 const Comment     = require('../models/comments.js');
+const User            = require('../models/users.js');
 
 // index route (GET)
 router.get('/', async (req, res) => {
-  try{
+  // try{
     const allTeapots = await Teapot.find();
-    if (req.session.logged) {
       res.render('./teapots/index.ejs', {
         allTeapots,
         username: req.session.username
       });
-      // res.send({allTeapots});
-      // res.send("Welcome to the Teapot Wonder Dome Index")
-    } else {
-      res.redirect('/user/login');
-    }
-  } catch (err) {
-    res.send(err.message);
-  }
+  // } catch (err) {
+  //   res.send(err.message);
+  // }
 });
 
 // new  (GET)
-router.get('/new', (req, res) => {
-  try {
-    // res.send("This is the route to add a NEW teapot");
-    res.render('./teapots/new.ejs');
-  } catch (err) {
-    res.send(err.message);
+router.get('/new', async (req, res) => {
+  if (req.session.logged){
+    try {
+      const currentUser = await User.find({username: req.session.username});
+      res.render('./teapots/new.ejs', {currentUser});
+    } catch (err) {
+      res.send(err.message);
+    }
+  } else {
+    try {
+      res.redirect('/user/login');
+    } catch (err) {
+      res.send(err.message);
+    }
   }
 });
 
 // create (POST)
 router.post('/', async (req, res) => {
-  try {
-    const newTeapot = await Teapot.create(req.body);
-    res.redirect('/teapots/'+newTeapot.id);
-    // res.redirect('/teapots/');
-  } catch (err) {
-    res.send(err.message);
+  if (req.session.logged) {
+    try {
+      const newTeapot = await Teapot.create(req.body);
+      res.redirect('/teapots/'+newTeapot.id);
+    } catch (err) {
+      res.send(err.message);
+    }
+  } else {
+    try {
+      res.redirect('/user/login');
+    } catch (err) {
+      res.send(err.message);
+    }
   }
 });
 
@@ -51,8 +61,8 @@ router.get('/:id', async (req, res) => {
   try {
     const oneTeapot = await Teapot.findById(req.params.id);
     const comments  = await Comment.find( {teapot: oneTeapot._id});
-    // res.send({oneTeapot});
-    res.render('./teapots/show.ejs', {oneTeapot, comments});
+    res.render('./teapots/show.ejs', {
+      oneTeapot, comments, username: req.session.username});
   } catch (err) {
     res.send(err.message);
   }
@@ -60,33 +70,50 @@ router.get('/:id', async (req, res) => {
 
 // edit (GET)
 router.get('/:id/edit', async (req, res) => {
-  try {
-    const editThisTeapot = await Teapot.findById(req.params.id);
-    res.render('./teapots/edit.ejs', {editThisTeapot});
-    // res.send({editThisTeapot});
-  } catch (err) {
-    res.send(err.message);
+  if (req.session.logged) {
+    try {
+      const editThisTeapot = await Teapot.findById(req.params.id);
+      // if (req.session.username === editThisTeapot.submitted_by){
+        res.render('./teapots/edit.ejs', {
+          editThisTeapot, username: req.session.username
+        });
+      // } else {
+      //   res.render('./sessions/wrong-user.ejs');
+      // }
+    } catch (err) {
+      res.send(err.message);
+    }
+  } else {
+    res.redirect('/user/login');
   }
 });
 
 // update (PUT)     --------->> teapots only <<--------------
 router.put('/:id', async (req, res) => {
-  try {
-    const updatedTeapot = await Teapot.findByIdAndUpdate(req.params.id, req.body);
-    res.redirect('/');
-  } catch (err) {
-    res.send(err.message);
+  if (req.session.logged){
+    try {
+      const updatedTeapot = await Teapot.findByIdAndUpdate(req.params.id, req.body);
+      res.redirect('/');
+    } catch (err) {
+      res.send(err.message);
+    }
+  } else {
+    res.redirect('/user/login');
   }
 });
 
 // delete (DELETE) ----->> only deletes teapots now <<--------
 router.delete('/:id', async (req, res) => {
-  try {
-    const teapot = await Teapot.findByIdAndRemove(req.params.id);
-    await Comment.remove({ teapot: teapot._id});
-    res.redirect('/');
-  } catch (err) {
-    res.send(err.message);
+  if (req.session.logged) {
+    try {
+      const teapot = await Teapot.findByIdAndRemove(req.params.id);
+      await Comment.remove({ teapot: teapot._id});
+      res.redirect('/');
+    } catch (err) {
+      res.send(err.message);
+    }
+  } else {
+    res.redirect('/user/login');
   }
 });
